Broke tests apart and added fixtures
1. Broke the main `build` test up into tests for individual asset types
2. Added fixtures directory to house small apps for testing individual
   bits of functionality, rather than adding more and more to the main
   example app
3. Added some helpers:
   * Build helpers to run some common functions between each of the
     asset build tests
   * Runners to run the `build` and `serve` tasks

Note that the `build` and `serve` runner helpers are placed in their
own file so that they can be launched in their own forked process.
The problem with running them inline is that they need to be run
from each fixture app's root directory, but their async nature makes
it impossible to know which dir you're in when the task actually fires.
By running them in their own process, we can `chdir` to the right directory,
then fork the process, which stays in that directory no matter where the
main process `chdir`s to next.